requirements.txt
redis>=5.0.0
chromadb-client>=0.5.0
psutil>=6.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
